# docs/openapi-phase4.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DreamContract Phase 4 API — Blocks 1–3",
    "version": "0.3.0",
    "description": "Incremental contract for implemented Phase 4 reads only. Existing Phase 2/3 endpoints retain their contracts. Secure mode (AUTH_ENABLED=1) is required for beta identity. Later journey endpoints are not advertised as callable until implemented."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
    "/dashboard": {
      "get": {
        "operationId": "getDashboard",
        "summary": "getDashboard",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "webSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Read-only projection of current persisted state.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or session revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Journey eligibility requirement not met.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Method not supported.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error; reference is safe to report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/journey/status": {
      "get": {
        "operationId": "getJourneyStatus",
        "summary": "getJourneyStatus",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "webSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Read-only projection of current persisted state.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or session revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Journey eligibility requirement not met.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Method not supported.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error; reference is safe to report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/guidance": {
      "get": {
        "operationId": "getGuidance",
        "summary": "getGuidance",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "webSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Read-only projection of current persisted state.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidanceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or session revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Journey eligibility requirement not met.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Method not supported.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error; reference is safe to report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/week": {
      "get": {
        "operationId": "getWeek",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "webSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation. Preparation is idempotent; repeated identical decisions return replayed=true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeekResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Verification or stage restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, not owned or unrevealed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Closed window, unavailable match or conflicting decision.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "JSON body required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error; diagnostic reference returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/week/prepare": {
      "post": {
        "operationId": "prepareWeek",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "webSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation. Preparation is idempotent; repeated identical decisions return replayed=true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeekResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Verification or stage restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, not owned or unrevealed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Closed window, unavailable match or conflicting decision.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "JSON body required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error; diagnostic reference returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "required": [],
                "additionalProperties": false
              }
            }
          }
        },
        "description": "Cookie-authenticated mutations require a matching Origin/Referer. Native bearer requests do not use cookie CSRF. No client clock or actor ID is accepted."
      }
    },
    "/matches/{match_id}": {
      "get": {
        "operationId": "getMatch",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "webSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation. Preparation is idempotent; repeated identical decisions return replayed=true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Verification or stage restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, not owned or unrevealed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Closed window, unavailable match or conflicting decision.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "JSON body required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error; diagnostic reference returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/matches/{match_id}/actions": {
      "post": {
        "operationId": "decideMatch",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "webSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation. Preparation is idempotent; repeated identical decisions return replayed=true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Verification or stage restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, not owned or unrevealed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Closed window, unavailable match or conflicting decision.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "JSON body required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error; diagnostic reference returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionInput"
              }
            }
          }
        },
        "description": "Cookie-authenticated mutations require a matching Origin/Referer. Native bearer requests do not use cookie CSRF. No client clock or actor ID is accepted."
      }
    },
    "/lock-ins/current": {
      "get": {
        "operationId": "getCurrentLockIn",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "webSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation. Preparation is idempotent; repeated identical decisions return replayed=true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentLockInResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Verification or stage restriction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, not owned or unrevealed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Closed window, unavailable match or conflicting decision.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "JSON body required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error; diagnostic reference returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      },
      "webSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dhashu_auth"
      }
    },
    "schemas": {
...
      "Guidance": {
        "type": "object",
        "properties": {
          "headline": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "cta": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "destination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Surface"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "headline",
          "body",
          "cta",
          "destination"
        ],
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "data": {
            "type": "null"
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "detail": {},
              "reference": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "code",
              "message",
              "detail",
              "reference"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "data",
          "error"
        ],
        "additionalProperties": false
      },
      "JourneyStatusResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JourneyStatus"
          },
          "error": {
            "type": "null"
          }
        },
        "required": [
          "data",
          "error"
        ],
        "additionalProperties": false
      },
      "GuidanceResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Guidance"
          },
          "error": {
            "type": "null"
          }
        },
        "required": [
          "data",
          "error"
        ],
        "additionalProperties": false
      },
      "Match": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "week": {
            "type": "integer"
          },
          "slot": {
            "type": "integer"
          },
          "revealed_at": {
            "type": "string"
          },
          "window_closes_at": {
            "type": "string"
          },
          "action": {
            "enum": [
              "none",
              "interest",
              "pass"
            ]
          },
          "status": {
            "enum": [
              "not_yet_revealed",
              "open",
              "acted",
              "no_response",
              "closed"
            ]
          },
          "candidate": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "display_name": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "visions": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "bgv_status": {
                    "type": "string"
                  },
                  "stats": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "age": {},
                      "height_cm": {},
                      "profession": {},
                      "income_band": {},
                      "education": {},
                      "diet": {},
                      "nationality": {}
                    }
                  }
                },
                "required": [
                  "display_name",
                  "city",
                  "visions",
                  "bgv_status",
                  "stats"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ]
          },
          "their_interest": {
            "type": "boolean"
          },
          "allowed_actions": {
            "type": "array",
            "items": {
              "enum": [
                "interest",
                "pass"
              ]
            }
          }
        },
        "required": [
          "id",
          "week",
          "slot",
          "revealed_at",
          "window_closes_at",
          "action",
          "status",
          "candidate",
          "their_interest",
          "allowed_actions"
        ],
        "additionalProperties": false
      },
...
      "Decision": {
        "type": "object",
        "properties": {
          "match_id": {
            "type": "string"
          },
          "action": {
            "enum": [
              "interest",
              "pass"
            ]
          },
          "replayed": {
            "type": "boolean"
          }
        },
        "required": [
          "match_id",
          "action",
          "replayed"
        ],
        "additionalProperties": false
      },
      "DecisionInput": {
        "type": "object",
        "required": [
          "action"
        ],
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "interest",
              "pass"
            ]
          },
          "pass_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "WeekResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Week"
          },
          "error": {
            "type": "null"
          }
        },
        "required": [
          "data",
          "error"
        ],
        "additionalProperties": false
      },
      "MatchResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Match"
          },
          "error": {
            "type": "null"
          }
        },
        "required": [
          "data",
          "error"
        ],
        "additionalProperties": false
      },
      "CurrentLockInResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CurrentLockIn"
          },
          "error": {
            "type": "null"
          }
        },
        "required": [
          "data",
          "error"
        ],
        "additionalProperties": false
      },
      "DecisionResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Decision"
          },
          "error": {
            "type": "null"
          }
        },
        "required": [
          "data",
          "error"
        ],
        "additionalProperties": false
      },
...
    }
  }
}